+++ NEWS.rst
@@ -1,3 +1,11 @@
+New in Elpy 1.22.0
+==================
+
+- Elpy now supports the Black source code formatter.
+- Make Elpy work more nicely with the new Flymake.
+- Support Emacs 26.1
+- Lots of bugfixes.
+
 New in Elpy 1.21.0
 ==================
 
@@ -5,7 +13,6 @@ New in Elpy 1.21.0
 - Elpy now knows how to deal with the flymake in the upcoming Emacs
   release.
 - Some minor improvements in error messages and warnings.
-  
 
 New in Elpy 1.20.0
 ==================
@@ -32,7 +39,7 @@ Thanks to Craig MacEachern, Daniel Gopar, Gaby Launay, Maxim
 Cournoyer, Nicholas D. Steeves and Sam Steingold for their help in
 making this release!
 
-  
+
 New in Elpy 1.18.0
 ==================
 
@@ -62,7 +69,7 @@ New in Elpy 1.18.0
 We are happy to report that Elpy now has more maintainers! Daniel
 Gopar, Rainer Gemulla and @galaunay are now helping regularly with
 this project.
-  
+
 Thanks to all the contributors!